Rpc module refactoring
