+++ src/lib.rs
@@ -4,7 +4,6 @@ pub mod filter;
 pub mod metrics;
 pub mod pubsub;
 pub mod rpc;
-pub mod rpc_monitor;
 pub mod types;
 
 const DEFAULT_GAI_QUEUE_SIZE: usize = 2 << 19;
